refactor comressed_posting
