+++ lib/segment/src/index/field_index/full_text_index/compressed_posting/compressed_common.rs
