+++ lib/segment/src/index/field_index/full_text_index/compressed_posting/mod.rs
@@ -1,0 +1,4 @@
+pub mod compressed_posting_list;
+mod compressed_chunks_reader;
+pub mod compressed_common;
+pub mod compressed_posting_visitor;
