+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -1,12 +1,9 @@
-mod compressed_posting_list;
 mod inverted_index;
 mod posting_list;
 mod postings_iterator;
 pub mod text_index;
 mod tokenizers;
 
-mod compressed_chunks_reader;
-pub mod compressed_common;
-mod compressed_posting_visitor;
 #[cfg(test)]
 mod tests;
+mod compressed_posting;
