+++ lib/segment/src/index/field_index/full_text_index/postings_iterator.rs
@@ -1,8 +1,8 @@
 use common::types::PointOffsetType;
 
 use super::posting_list::PostingList;
-use crate::index::field_index::full_text_index::compressed_posting_list::CompressedPostingList;
-use crate::index::field_index::full_text_index::compressed_posting_visitor::CompressedPostingVisitor;
+use crate::index::field_index::full_text_index::compressed_posting::compressed_posting_list::CompressedPostingList;
+use crate::index::field_index::full_text_index::compressed_posting::compressed_posting_visitor::CompressedPostingVisitor;
 
 pub fn intersect_postings_iterator<'a>(
     mut postings: Vec<&'a PostingList>,
